add -> login service and admin seed

+++ src/controllers/user_controller.py
@@ -5,11 +5,13 @@ from src.models.user_model import (
     Message, UserUpdate
 )
 from src.utils.responses import responses
-from src.utils.types import T_Session
+from src.utils.types import T_Session, T_CurrentUser
 from src.service.user_service import (
     create_user_service, get_all_users_service, get_user_by_id_service,
     delete_user_by_id_service, update_user_service
 )
+from src.utils.validations import authorize_user, is_admin
+from src.models.exceptions import ForbiddenException
 
 
 router = APIRouter(prefix='/user', tags=['users'])
@@ -36,15 +38,17 @@ def create_user(user: User, session: T_Session) -> UserResponse:
     responses={
         **responses['bad_request'],
         **responses['internal_server_error'],
-        #**responses['unauthorized'],
-        #**responses['forbidden']
+        **responses['unauthorized'],
+        **responses['forbidden']
     }
 )
 def get_users(
     session: T_Session, 
-    #current_user: T_CurrentUser,
+    current_user: T_CurrentUser,
     offset: int = 0, limit: int = 100
 ) -> list[UserResponse]:
+    if not is_admin(current_user, session):
+        raise ForbiddenException(detail='User not allowed to access this resource')
     users = get_all_users_service(offset, limit, session)
     return {'users': users}
 
@@ -56,15 +60,16 @@ def get_users(
     responses={
         **responses['bad_request'],
         **responses['internal_server_error'],
-        #**responses['unauthorized'],
-        #**responses['forbidden']
+        **responses['unauthorized'],
+        **responses['forbidden']
     }   
 )
 def get_user_by_id(
     user_id: int, session: T_Session, 
-    #current_user: T_CurrentUser
+    current_user: T_CurrentUser
 ) -> UserResponse:
-    #authorize_user(current_user.id, user_id)
+    if not is_admin(current_user, session):
+        authorize_user(current_user.id, user_id)
     return get_user_by_id_service(user_id, session)
 
 
@@ -81,9 +86,14 @@ def get_user_by_id(
 )
 def delete_user_by_id(
     user_id: int, session: T_Session, 
-    #current_user: T_CurrentUser
+    current_user: T_CurrentUser
 ) -> Message:
-    #authorize_user(current_user.id, user_id)
+    user_is_admin = is_admin(current_user, session)
+    if not user_is_admin:
+        authorize_user(current_user.id, user_id)
+    user = get_user_by_id_service(user_id, session)
+    if user_is_admin and user.id == current_user.id:
+        raise ForbiddenException(detail='It is not possible to delete the admin user.')
     return delete_user_by_id_service(user_id, session)
 
 
@@ -101,7 +111,8 @@ def delete_user_by_id(
 def update_user(
     user_id: int, user_data: UserUpdate, 
     session: T_Session, 
-    #current_user: T_CurrentUser
+    current_user: T_CurrentUser
 ) -> UserResponse:
-    #authorize_user(current_user.id, user_id)
+    if not is_admin(current_user, session):
+        authorize_user(current_user.id, user_id)
     return update_user_service(user_id, user_data, session)
